[Chore] #66 - Category와 관련된 패키지 이동
