+++ src/main/java/com/arabook/arabook/category/repository/MainCategoryRepository.java
@@ -1,8 +1,8 @@
-package com.arabook.arabook.storage.domain.category.repository;
+package com.arabook.arabook.category.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.arabook.arabook.storage.domain.category.entity.MainCategory;
+import com.arabook.arabook.category.entity.MainCategory;
 
 public interface MainCategoryRepository
     extends JpaRepository<MainCategory, Long>, MainCategoryCustomRepository {}
