+++ src/main/java/com/arabook/arabook/category/repository/SubCategoryRepository.java
@@ -1,8 +1,8 @@
-package com.arabook.arabook.storage.domain.category.repository;
+package com.arabook.arabook.category.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.arabook.arabook.storage.domain.category.entity.SubCategory;
+import com.arabook.arabook.category.entity.SubCategory;
 
 public interface SubCategoryRepository
     extends JpaRepository<SubCategory, Long>, SubCategoryCustomRepository {}
